feat: AppContext 분리 및 폴더 구조 변경
