+++ src/contexts/AppContext.tsx
@@ -1,0 +1,4 @@
+import { createContext } from "react";
+import { AppContextType } from "./type";
+
+export const AppContext = createContext<AppContextType | undefined>(undefined);
