+++ src/contexts/useAppContext.ts
@@ -1,0 +1,13 @@
+import { useContext } from "react";
+import { AppContext } from "./AppContext";
+
+export function useAppContext() {
+  const context = useContext(AppContext);
+  console.log(context);
+
+  if (context === undefined) {
+    throw new Error("useApp must be used within an AppProvider");
+  }
+
+  return context;
+}
